Move code for inferring build targets to a separate file
